Implement Corpus IR analysis engine (CIAN001A) — Score to WorkAnalysis
Adds analytical decomposition across seven domains from Score IR content:

Harmonic: chord vocabulary from HarmonicAnnotationLayer or on-the-fly
recognition via recognize_chord/chord_to_numeral, progression bigram
cataloguing, harmonic rhythm profile, tonal plan from key map.

Melodic: per-voice range, interval distribution, conjunct proportion,
chromaticism rate via compute_melody_statistics from MLCT001A.

Rhythmic: duration distribution, onset density per bar, tempo profile,
rest proportion, metrical complexity, note density by section.

Formal: section plan from SectionMap, form classification (binary,
ternary, rondo) via inline pattern matching, section proportions with
golden ratio proximity, tonal plan.

Textural: per-bar density curve, average register span, section density.

Dynamic: dynamic distribution, range, hairpin count, subito dynamics,
dynamic change rate.

Orchestration: instrument usage proportion, melody carrier distribution
(multi-part scores only; nullopt for single-part).

wf_analyze_work now accepts an optional Score pointer; when provided,
calls analyze_score to populate all analysis fields before marking
analysis_complete. Existing callers (no Score) retain prior behaviour.

CadenceEvent.type changed from enum to string to resolve ODR violation
between HRCD001A.h and CITP001A.h CadenceType definitions.

16 new tests (57 assertions), 1493 total passing.

+++ src/Sunny.Core/Corpus/CISZ001A.cpp
@@ -143,7 +143,7 @@ json harmonic_analysis_j(const HarmonicAnalysisRecord& h) {
     for (const auto& c : h.cadence_inventory) {
         cads.push_back({
             {"position", score_time_j(c.position)},
-            {"type", static_cast<int>(c.type)},
+            {"type", c.type},
             {"approach", c.approach},
             {"section_context", c.section_context},
             {"is_structural", c.is_structural}
